+++ api/fetcher.py
@@ -3,32 +3,98 @@ import requests
 import urllib.parse
 from datetime import datetime
 from bs4 import BeautifulSoup as bs
+from selenium import webdriver
+from selenium.webdriver.chrome.service import Service
+from selenium.webdriver.common.by import By
+from selenium.webdriver.chrome.options import Options
+from pyvirtualdisplay import Display
 
 log = logging.getLogger("api.fetcher")
-import configparser
+
+
+class WetterComFetcher:
+    """
+    Wetter.com dynamically updates the value for the current temperature.
+    Usually, it first displays a static temperature and then reloads it with a more recent (dynamic) value.
+
+    This class provides a method for catching the static value and 
+    another for catching the dynamic value.
+    """
+
+    @staticmethod
+    def get_data_static(url):
+        """
+        Fetches the static temperature data from Wetter.com link for a city/region
+        """
+        try:
+            response = requests.get(url)
+        except requests.exceptions.ConnectionError as e:
+            log.error("Wetter.com connection problem: " + str(e))
+            return None
+
+        if response.status_code == 200:
+            soup = bs(response.text, 'html.parser')
+            temperature_element = soup.find('div', class_='delta rtw_temp')
+            if temperature_element:
+                temperature = int(temperature_element.text.strip().replace("°", "").replace("C", ""))
+                return temperature
+            else:
+                log.error("Temperature element not found on the page.")
+                return None
+        else:
+            log.error("Failed to retrieve weather data.")
+            return None
+
+    @staticmethod
+    def get_data_dynamic(url):
+        """
+
+        Fetches the dynamic temperature data from Wetter.com link for a city/region
+        """
+        display = Display(visible=0, size=(1600, 1200))
+        display.start()
+        options = Options()
+        options.add_argument('--disable-blink-features=AutomationControlled')
+        service = webdriver.ChromeService(executable_path='/usr/lib/chromium-browser/chromedriver')
+        driver = webdriver.Chrome(service=service, options=options)
+        try:
+            driver.get(url)
+            found_temp = driver.find_element(By.XPATH, '//div[@class="delta rtw_temp"]')
+            return int(found_temp.text.replace('°C', ''))
+
+        except Exception as e:
+            log.error(f"An error occurred while dynamically fetching temperature data: {str(e)}")
+            return None
+        finally:
+            driver.quit()
 
 
 class GoogleFetcher:
+
     @staticmethod
     def get_weather_data(location: str):
-        URL = "https://www.google.com/search?lr=lang_en&ie=UTF-8&q=weather" + location
-        USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/94.0.4606.81 Safari/537.36"
-        LANGUAGE = "en-US,en;q=0.5"
+        """
+        Fetches data from Google Weather for a specific location.
+
+        :location: e.g. "New York"
+        """
+
+        url = "https://www.google.com/search?lr=lang_en&ie=UTF-8&q=weather" + location
+        user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/94.0.4606.81 Safari/537.36"
+        language = "en-US,en;q=0.5"
         try:
             session = requests.Session()
-            session.headers["User-Agent"] = USER_AGENT
-            session.headers["Accept-Language"] = LANGUAGE
-            session.headers["Content-Language"] = LANGUAGE
-            html = session.get(URL)
+            session.headers["User-Agent"] = user_agent
+            session.headers["Accept-Language"] = language
+            session.headers["Content-Language"] = language
+            html = session.get(url)
             soup = bs(html.text, "html.parser")
 
-            result = {}
-            result["region"] = soup.find("div", attrs={"id": "wob_loc"}).text
-            result["temp_now"] = float(soup.find("span", attrs={"id": "wob_tm"}).text)
-            result["precipitation"] = float(soup.find("span", attrs={"id": "wob_pp"}).text.replace("%", ""))
-            result["humidity"] = float(soup.find("span", attrs={"id": "wob_hm"}).text.replace("%", ""))
-            result["wind"] = float(soup.find("span", attrs={"id": "wob_ws"}).text.replace(" km/h", ""))
-            return result
+            return {"region": soup.find("div", attrs={"id": "wob_loc"}).text,
+                    "temp_now": float(soup.find("span", attrs={"id": "wob_tm"}).text),
+                    "precipitation": float(soup.find("span", attrs={"id": "wob_pp"}).text.replace("%", "")),
+                    "humidity": float(soup.find("span", attrs={"id": "wob_hm"}).text.replace("%", "")),
+                    "wind": float(soup.find("span", attrs={"id": "wob_ws"}).text.replace(" km/h", ""))}
         except requests.exceptions.ConnectionError as e:
             log.error("Google Weather connection problem: " + str(e))
             return None
@@ -66,7 +132,7 @@ class Fetcher:
     def _handle_ok_status_code(self, response):
         return response.json()
 
-    def _handle_bad_status_code(self, code: int) -> int:
+    def _handle_bad_status_code(self, code):
         log.error(f"Error: Failed to fetch weather data. Status code: {code}")
         return None